Refactor foundation modules into organized structure
## Phase 1: Foundation Modules (Zero Dependencies)

### New Structure
- emdx/config/settings.py (from config.py)
- emdx/utils/git.py (from utils.py)
- emdx/display/formatting.py (from tag_display.py)

### Updated Imports
- migrations.py: from .config.settings import get_db_path
- core.py: from emdx.utils.git import get_git_project
- core.py, tag_commands.py, textual_browser_minimal.py: from emdx.display.formatting import format_tags

### Benefits
✅ Better organization and modularity
✅ Clear separation of concerns
✅ Foundation for future refactoring phases
✅ Zero functionality changes - all tests pass

Implements emdx #525 foundation refactoring gameplan.
