+++ emdx/tag_commands.py
@@ -6,7 +6,7 @@ from rich.console import Console
 from rich.table import Table
 
 from emdx.sqlite_database import db
-from emdx.tag_display import format_tags
+from emdx.display.formatting import format_tags
 from emdx.tags import (
     add_tags_to_document,
     get_document_tags,
